# swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample recipes API. You can find out more about the API at https://github.com/PacktPublishing/Building-Distributed-Applications-in-Gin.",
        "title": "Recipes API",
        "contact": {
            "name": "Mohamed Labouardy",
            "url": "https://.com",
            "email": ".com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/recipes": {
            "get": {
                "description": "Returns list of recipes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "operationId": "listRecipes",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "description": "Create a new recipe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "operationId": "newRecipe",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/recipes/search": {
            "get": {
                "description": "Search recipes based on tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "operationId": "findRecipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recipe tag",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/recipes/{id}": {
            "put": {
                "description": "Update an existing recipe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "operationId": "updateRecipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the recipe",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Invalid recipe ID"
                    }
                }
            },
            "delete": {
                "description": "Delete an existing recipe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "operationId": "deleteRecipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the recipe",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Invalid recipe ID"
                    }
                }
            }
        }
    }
}